+++ cpp-terminal/platforms/conversion.hpp
