+++ cpp-terminal/platforms/macros.hpp
