+++ cpp-terminal/platforms/terminal.cpp
