+++ cpp-terminal/prompt.cpp
@@ -1,7 +1,7 @@
 #include "cpp-terminal/prompt.hpp"
 
 #include "cpp-terminal/input.hpp"
-#include "cpp-terminal/private/conversion.hpp"
+#include "cpp-terminal/platforms/conversion.hpp"
 #include "cpp-terminal/terminal.hpp"
 #include "cpp-terminal/tty.hpp"
